\name{print.traps}

\alias{print.traps}

\title{ Print Detectors }

\description{
Print method for \code{traps} objects.
}

\usage{
\method{print}{traps}(x, \dots)
}

\arguments{
  \item{x}{ \code{traps} object }
  \item{\dots}{ arguments to pass to \code{\link{print.default}} }
}

\seealso{ \code{\link{print}}, \code{\link{traps}} }

\examples{
print(make.grid(nx = 5, ny = 3))
}


