\name{secr-deprecated}
\alias{secr-deprecated}

\title{Deprecated Functions in Package \pkg{secr}}

\description{


  These functions will be removed from future versions of \pkg{secr}.

}

\usage{

# Deprecated in 4.4.6

# None

}

\seealso{

\code{\link{secr-defunct}}

}

