\name{Miscellaneous}

\alias{gethazard}
\alias{make.lookup}

\title{Miscellaneous Internal Functions}

\description{

These functions are exported for possible use elsewhere. They include wrappers for 
C++ code that may be useful in other packages. They are not intended to be 
called by most users.

}

\usage{

gethazard(m, binomN, cc, hk, PIA, usge)

reparameterize(realparval, detectfn, details, mask, traps, D, s)

}

\arguments{

  \item{m}{}
  \item{binomN}{}
  \item{cc}{}
  \item{hk}{}
  \item{PIA}{}
  \item{usge}{}
  
  \item{realparval}{}
  \item{detectfn}{}
  \item{details}{}
  \item{mask}{}
  \item{traps}{}
  \item{D}{}
  \item{s}{}

}

\details{

Not for general use. Refer to secr code.

}

\value{

gethazard

reparameterize

}

