\name{Internal}

\alias{boundarytoSF}
\alias{Fletcher.chat}

\title{Internal Functions}

\description{ 

Functions called internally by \pkg{secr} and not exported.

}

\usage{

boundarytoSF (poly)

Fletcher.chat (observed, expected, np, verbose = TRUE, 
    type = c('Fletcher', 'Pearson', 'both'))

}

\arguments{

  \item{poly}{data to define one or more polygons}

  \item{observed}{integer vector of observed counts}
  \item{expected}{numeric vector of expected counts}
  \item{np}{integer number of parameters estimated}
  \item{verbose}{logical; if TRUE returns extended output}
  \item{type}{character}

}

\details{

  The function \code{boundarytoSF} converts various 
  possible polygon input formats to a standard form (sfc).
  
  Possible inputs are:
  
  \tabular{lll}{
  Input \tab From \tab Note \cr
  2-column matrix or dataframe \tab base R\cr
  SpatialPolygons \tab \pkg{sp} \cr
  SpatialPolygonsDataFrame \tab \pkg{sp} \cr
  SpatVector \tab \pkg{terra} \cr
  sf \tab \pkg{sf} \tab geometry type POLYGON or MULTIPOLYGON \cr
  sfc \tab \pkg{sf} \tab geometry type POLYGON or MULTIPOLYGON \cr
  }
  
  Matrix input defines a single polygon.

\code{Fletcher.chat} applies the overdispersion formula of Fletcher (2012) or computes the conventional variance inflation factor \eqn{X^2/df}. It is used by \code{\link{chat.nk}} and \code{\link{adjustVarD}}. The inputs `observed' and `expected' are vectors of counts (e.g., number of distinct individuals per detector); `observed' may also be a list of such vectors, possibly simulated.

  }

\value{

\code{boundarytoSF} -- Spatial object of \pkg{sf} class sfc, containing 
a geometry set of type POLYGON or MULTIPOLYGON. NULL input results in NULL output.

\code{Fletcher.chat} -- Output depends on `verbose', `observed' and `type'. 

If `observed' is a list of nk vectors (usually generated by simulation) then regardless of `verbose' the output is a vector of (Fletcher or Pearson) \eqn{\hat c} values, one element for each component of `observed', unless type = "both" when the output is a list of two such vectors. 

If `observed' is a simple vector then `Verbose' output is a list comprising input values, various summary statistics, and the computed Fletcher overdispersion (`chat'). The statistic `cX2' is the conventional (Pearson) variance inflation factor \eqn{X^2/df}. 

For verbose = FALSE, a single estimate of \eqn{\hat c} is returned when \code{type = "Fletcher"} or \code{type = "Pearson"}, otherwise a vector of the two estimates.

}

\seealso{

\code{\link{pointsInPolygon}}, \href{https://www.otago.ac.nz/density/pdfs/secr-spatialdata.pdf}{secr-spatialdata.pdf},

\code{\link{chat.nk}}

}

\references{

Fletcher, D. (2012) Estimating overdispersion when fitting a generalized linear model to sparse data.
\emph{Biometrika} \bold{99}, 230--237.

Hijmans, R. J. (2022) terra: Spatial Data Analysis. R package version 1.5-14. https://rspatial.org/terra/

Pebesma, E. (2018) Simple features for R: standardized support for spatial vector data. \emph{The R Journal} \bold{10(1)}, 439--446.
  https://doi.org/10.32614/RJ-2018-009

Pebesma, E.J. and Bivand, R. S. (2005) Classes and methods for spatial data in R. \emph{R News} \bold{5(2)}, 9--13.
  https://cran.r-project.org/doc/Rnews/Rnews_2005-2.pdf.

}

\examples{

\dontrun{

poly <- cbind(x = c(0,6,6,0,0), y = c(0,0,6,6,0))  
secr:::boundarytoSF(poly)

}

}